support for very large maps
git-svn-id: https://svn.osgeo.org/grass/grass-addons@43678 15284696-431f-4ddb-bdfa-cd5b030d7da7

+++ grass7/raster/r.clump2/local_proto.h
@@ -28,7 +28,7 @@ extern long pqsize;
 extern CELL *clump_id;
 
 /* pq.c */
-int init_pq(int);
+int init_pq(long);
 int free_pq(void);
 int add_pnt(long);
 long drop_pnt(void);
